[#21] Controller to Kafka
- :mailbox: producing data to Kafka
- :100: coverage
