+++ src/main/java/nl/nn/app/player/stream/PlayerStreamProducer.java
@@ -24,6 +24,7 @@ import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import nl.nn.app.player.view.PlayerVO;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.kafka.support.SendResult;
 import org.springframework.stereotype.Component;
@@ -33,12 +34,13 @@ import org.springframework.util.concurrent.ListenableFuture;
 @Slf4j
 @RequiredArgsConstructor
 public class PlayerStreamProducer {
-    public static final String TOPIC = "nn-player";
+    @Value("${spring.kafka.properties.topic.player}")
+    public String topic;
 
     private final KafkaTemplate<UUID, PlayerVO> kafka;
 
     public ListenableFuture<SendResult<UUID, PlayerVO>> send(final PlayerVO data) {
-        log.debug("produce: topic='{}' key='{}' data='{}'", TOPIC, data.getId(), data);
-        return kafka.send(TOPIC, data.getId(), data);
+        log.debug("produce: topic='{}' key='{}' data='{}'", topic, data.getId(), data);
+        return kafka.send(topic, data.getId(), data);
     }
 }
