+++ src/main/resources/application.properties
@@ -25,3 +25,6 @@ management.endpoints.web.exposure.include=health,info,mappings,metrics
 # Kafka Configs
 spring.kafka.producer.key-serializer=org.springframework.kafka.support.serializer.ToStringSerializer
 spring.kafka.producer.value-serializer=org.springframework.kafka.support.serializer.JsonSerializer
+spring.kafka.properties.topic.player=nn-player
+spring.kafka.properties.topic.item=nn-item
+spring.kafka.properties.topic.purchase=nn-purchase
