+++ src/test/resources/application.properties
@@ -25,3 +25,7 @@ spring.kafka.consumer.auto-offset-reset=earliest
 spring.kafka.consumer.key-deserializer=org.apache.kafka.common.serialization.UUIDDeserializer
 spring.kafka.consumer.value-deserializer=org.springframework.kafka.support.serializer.JsonDeserializer
 spring.kafka.consumer.properties.spring.json.trusted.packages=nl.nn.app.*
+spring.kafka.properties.topic.player=nn-player
+spring.kafka.properties.topic.item=nn-item
+spring.kafka.properties.topic.purchase=nn-purchase
+
